refactor: split tauri runtime helpers
