+++ desktop/src-tauri/src/runtime/mod.rs
@@ -3,4 +3,5 @@ pub(crate) mod profile;
 pub(crate) mod provider;
 pub(crate) mod proxy;
 pub(crate) mod science;
+pub(crate) mod system;
 pub(crate) mod transaction;
